+++ src/ReaderEither.ts
@@ -178,9 +178,9 @@ export const orElse: <E1, R1, E2, B>(
  * @category combinators
  * @since 3.0.0
  */
-export const orElseFirst: <E1, R2, E2, B>(
-  onLeft: (e: E1) => ReaderEither<R2, E2, B>
-) => <R1, A>(ma: ReaderEither<R1, E1, A>) => ReaderEither<R1 & R2, E1 | E2, A> = /*#__PURE__*/ eitherT.orElseFirst(
+export const tapError: <E1, R2, E2, _>(
+  onLeft: (e: E1) => ReaderEither<R2, E2, _>
+) => <R1, A>(self: ReaderEither<R1, E1, A>) => ReaderEither<R1 & R2, E1 | E2, A> = /*#__PURE__*/ eitherT.tapError(
   reader.Monad
 )